# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "IAA 投放运营中台 API",
    "description": "IAA 投放运营中台 Java 业务服务契约；示例数据均为虚构。模型暂由 FastAPI 导出，x-implementation 表示 Java 状态；x-reference-implementation 保留 Python 对照服务的实际状态。",
    "version": "1.0.0-rc.1"
  },
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "认证与偏好"
        ],
        "summary": "账号密码登录",
        "description": "校验账号口令，签发短期访问令牌与刷新令牌。账号不存在、口令错误、账号停用及锁定统一返回 INVALID_CREDENTIALS。Java 业务服务按账号及来源 IP 限制失败登录，锁定期内正确口令也拒绝；成功登录或改密提交后清零账号与当前 IP 的失败计数。",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              },
              "examples": {
                "demo": {
                  "summary": "演示账号登录",
                  "value": {
                    "username": "demo.operator",
                    "password": "<密码>"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "登录成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPair"
                },
                "examples": {
                  "demo": {
                    "summary": "登录成功",
                    "value": {
                      "access_token": "<access-token>",
                      "refresh_token": "<refresh-token>",
                      "token_type": "bearer",
                      "expires_in": 900,
                      "must_change_password": false
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "未登录或令牌已失效",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "请求参数校验失败",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [],
        "x-requirements": [
          "IAA-REQ-001"
        ],
        "x-screens": [
          "login"
        ],
        "x-fields": [
          "username"
        ],
        "x-implementation": "implemented",
        "x-reference-implementation": {
          "runtime": "FastAPI",
          "status": "implemented"
        },
        "x-implementation-source": "backend-java/src/main/java/com/iaaops/iam/web/AuthController.java"
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "tags": [
          "认证与偏好"
        ],
        "summary": "刷新访问令牌",
        "description": "用刷新令牌换取新的令牌对，旧刷新令牌随即失效；已失效的刷新令牌再次出现时吊销该账号全部刷新令牌。",
        "operationId": "refreshToken",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              },
              "examples": {
                "demo": {
                  "summary": "换发令牌",
                  "value": {
                    "refresh_token": "<refresh-token>"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "已换发",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPair"
                },
                "examples": {
                  "demo": {
                    "summary": "换发成功",
                    "value": {
                      "access_token": "<access-token>",
                      "refresh_token": "<refresh-token>",
                      "token_type": "bearer",
                      "expires_in": 900,
                      "must_change_password": false
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "未登录或令牌已失效",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "请求参数校验失败",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [],
        "x-requirements": [
          "IAA-REQ-001"
        ],
        "x-screens": [],
        "x-fields": [],
        "x-implementation": "implemented",
        "x-reference-implementation": {
          "runtime": "FastAPI",
          "status": "implemented"
        },
        "x-implementation-source": "backend-java/src/main/java/com/iaaops/iam/web/AuthController.java"
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": [
          "认证与偏好"
        ],
        "summary": "退出登录",
        "description": "吊销提交的刷新令牌。访问令牌无状态，最长 15 分钟后自然失效，前端退出时应同时丢弃。",
        "operationId": "logout",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              },
              "examples": {
                "demo": {
                  "summary": "退出当前会话",
                  "value": {
                    "refresh_token": "<refresh-token>"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "已退出"
          },
          "422": {
            "description": "请求参数校验失败",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [],
        "x-requirements": [
          "IAA-REQ-001"
        ],
        "x-screens": [
          "app-shell"
        ],
        "x-fields": [],
        "x-implementation": "implemented",
        "x-reference-implementation": {
          "runtime": "FastAPI",
          "status": "implemented"
        },
        "x-implementation-source": "backend-java/src/main/java/com/iaaops/iam/web/AuthController.java"
      }
    },
    "/api/v1/auth/password": {
      "post": {
        "tags": [
          "认证与偏好"
        ],
        "summary": "修改口令",
        "description": "临时口令账号也可调用。成功后吊销该账号全部刷新令牌并签发新的令牌对。",
        "operationId": "changePassword",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordChange"
              },
              "examples": {
                "demo": {
                  "summary": "首次登录改密",
                  "value": {
                    "current_password": "<当前口令>",
                    "new_password": "<至少十位的新口令>"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "已修改",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPair"
                },
                "examples": {
                  "demo": {
                    "summary": "修改成功",
                    "value": {
                      "access_token": "<access-token>",
                      "refresh_token": "<refresh-token>",
                      "token_type": "bearer",
                      "expires_in": 900,
                      "must_change_password": false
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "未登录或令牌已失效",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "请求参数校验失败",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-requirements": [
          "IAA-REQ-001"
        ],
        "x-screens": [
          "login"
        ],
        "x-fields": [],
        "x-implementation": "implemented",
        "x-reference-implementation": {
          "runtime": "FastAPI",
          "status": "implemented"
        },
        "x-implementation-source": "backend-java/src/main/java/com/iaaops/iam/web/AuthController.java"
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "认证与偏好"
        ],
        "summary": "当前账号、权限与导航",
        "description": "返回角色、权限、数据范围以及按权限裁剪后的导航分组。临时口令账号也可调用。",
        "operationId": "getCurrentPrincipal",
        "responses": {
          "200": {
            "description": "当前账号",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Principal"
                },
                "examples": {
                  "demo": {
                    "summary": "运营账号",
                    "value": {
                      "user_id": "usr_01",
                      "username": "demo.operator",
                      "display_name": "演示运营",
                      "tenant_id": "tenant_demo",
                      "roles": [
                        "operator"
                      ],
                      "permissions": [
                        "dashboard.read",
                        "metrics.real"
                      ],
                      "data_scope": {
                        "products": [
                          "晴空天气",
                          "脑力答题王"
                        ]
                      },
                      "must_change_password": false,
                      "visible_metrics": [
                        "cost",
                        "revenue",
                        "impressions",
                        "clicks",
                        "launches",
                        "callbacks",
                        "conversions",
                        "roi",
                        "ctr",
                        "cpm",
                        "cpc",
                        "click_arpu",
                        "launch_rate",
                        "cvr",
                        "conversion_cost",
                        "callback_cost",
                        "launch_callback_rate",
                        "revenue_gap"
                      ],
                      "navigation": [
                        {
                          "title": "看盘分析",
                          "pages": [
                            {
                              "id": "report-aggregate",
                              "title": "聚合",
                              "route": "/analysis/aggregate"
                            },
                            {
                              "id": "report-daily",
                              "title": "分天明细",
                              "route": "/analysis/daily"
                            },
                            {
                              "id": "report-trend",
                              "title": "趋势图",
                              "route": "/analysis/trend"
                            },
                            {
                              "id": "roi-anomalies",
                              "title": "ROI 异常清单",
                              "route": "/analysis/roi-anomalies"
                            },
                            {
                              "id": "raw-detail",
                              "title": "原始明细",
                              "route": "/analysis/raw"
                            }
                          ]
                        },
                        {
                          "title": "系统管理",
                          "pages": [
                            {
                              "id": "metric-catalog",
                              "title": "指标与字段说明",
                              "route": "/system/metrics"
                            }
                          ]
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "未登录或令牌已失效",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-requirements": [
          "IAA-REQ-001"
        ],
        "x-screens": [
          "app-shell"
        ],
        "x-fields": [
          "username",
          "roles",
          "permissions",
          "data_scope"
        ],
        "x-implementation": "implemented",
        "x-reference-implementation": {
          "runtime": "FastAPI",
          "status": "implemented"
        },
        "x-implementation-source": "backend-java/src/main/java/com/iaaops/iam/web/AuthController.java"
      }
    },
    "/api/v1/me/preferences": {
      "get": {
        "tags": [
          "认证与偏好"
        ],
        "summary": "读取账号界面偏好",
        "description": "主题模式、主题预设与各页面列设置按账号保存，换设备后保持一致。从未保存时返回默认值且 revision 为 0。",
        "operationId": "getPreferences",
        "responses": {
          "200": {
            "description": "当前偏好",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Preferences"
                },
                "examples": {
                  "demo": {
                    "summary": "深色极光蓝",
                    "value": {
                      "theme_mode": "dark",
                      "theme_preset": "aurora-blue",
                      "table_columns": {
                        "report-aggregate": [
                          "product",
                          "cost",
                          "revenue",
                          "roi",
                          "cpc",
                          "click_arpu"
                        ]
                      },
                      "revision": 4
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "未登录或令牌已失效",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "无权限，或超出账号的数据范围",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-requirements": [
          "IAA-REQ-004"
        ],
        "x-screens": [
          "app-shell"
        ],
        "x-fields": [
          "theme_mode",
          "theme_preset",
          "table_columns",
          "revision"
        ],
        "x-implementation": "implemented",
        "x-reference-implementation": {
          "runtime": "FastAPI",
          "status": "implemented"
        },
        "x-implementation-source": "backend-java/src/main/java/com/iaaops/iam/web/AuthController.java"
      },
      "put": {
        "tags": [
          "认证与偏好"
        ],
        "summary": "保存账号界面偏好",
        "description": "带上最后读到的 revision；期间被其他设备修改则返回 409。",
        "operationId": "updatePreferences",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreferencesUpdate"
              },
              "examples": {
                "demo": {
                  "summary": "保存列设置",
                  "value": {
                    "theme_mode": "dark",
                    "theme_preset": "aurora-blue",
                    "table_columns": {
                      "report-aggregate": [
                        "product",
                        "cost",
                        "revenue",
                        "roi",
                        "cpc",
                        "click_arpu"
                      ]
                    },
                    "revision": 3
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "已保存",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Preferences"
                },
                "examples": {
                  "demo": {
                    "summary": "保存后版本 +1",
                    "value": {
                      "theme_mode": "dark",
                      "theme_preset": "aurora-blue",
                      "table_columns": {
                        "report-aggregate": [
                          "product",
                          "cost",
                          "revenue",
                          "roi",
                          "cpc",
                          "click_arpu"
                        ]
                      },
                      "revision": 4
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "未登录或令牌已失效",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "无权限，或超出账号的数据范围",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "版本已变化，请刷新后重试",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "请求参数校验失败",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-requirements": [
          "IAA-REQ-004"
        ],
        "x-screens": [
          "app-shell"
        ],
        "x-fields": [
          "theme_mode",
          "theme_preset",
          "table_columns",
          "revision"
        ],
        "x-implementation": "implemented",
        "x-reference-implementation": {
          "runtime": "FastAPI",
          "status": "implemented"
        },
        "x-implementation-source": "backend-java/src/main/java/com/iaaops/iam/web/AuthController.java"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Error": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "default": "about:blank"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "integer",
            "title": "Status"
          },
          "code": {
            "type": "string",
            "enum": [
              "AUTH_REQUIRED",
              "INVALID_CREDENTIALS",
              "PASSWORD_CHANGE_REQUIRED",
              "FORBIDDEN",
              "NOT_FOUND",
              "METHOD_NOT_ALLOWED",
              "CONFLICT",
              "PAYLOAD_TOO_LARGE",
              "VALIDATION_FAILED",
              "RESULT_TOO_LARGE",
              "INTERNAL_ERROR",
              "NOT_IMPLEMENTED",
              "HTTP_ERROR"
            ],
            "title": "Code",
            "x-code-descriptions": {
              "AUTH_REQUIRED": "未登录、令牌缺失或已失效，需要重新登录。",
              "CONFLICT": "对象已被他人修改（revision 不一致）或唯一键冲突，刷新后重试。",
              "FORBIDDEN": "没有该操作的权限，或请求超出账号的数据范围。",
              "HTTP_ERROR": "其他 HTTP 错误，以 status 为准。",
              "INTERNAL_ERROR": "服务端异常，已记录，可稍后重试。",
              "INVALID_CREDENTIALS": "账号或口令错误，或登录暂被限制；不区分账号不存在、口令错误、账号停用与锁定。",
              "METHOD_NOT_ALLOWED": "路径存在，但不支持该 HTTP 方法。",
              "NOT_FOUND": "对象或路径不存在。",
              "NOT_IMPLEMENTED": "契约已确定，接口尚未实现。",
              "PASSWORD_CHANGE_REQUIRED": "正在使用临时口令，修改口令前只能访问当前账号与改密接口。",
              "PAYLOAD_TOO_LARGE": "上传文件超过大小限制。",
              "RESULT_TOO_LARGE": "分组数或导出行数超过上限，请缩小日期范围或减少分组维度。",
              "VALIDATION_FAILED": "请求参数未通过校验，errors 列出具体字段。"
            }
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FieldError"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errors"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title",
          "status",
          "code"
        ],
        "title": "Error",
        "description": "RFC 9457 Problem Details（application/problem+json），所有 4xx/5xx 统一使用。"
      },
      "FieldError": {
        "properties": {
          "field": {
            "type": "string",
            "title": "Field"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "field",
          "message"
        ],
        "title": "FieldError"
      },
...
      "LoginRequest": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Username"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Password"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "LoginRequest"
      },
...
      "PasswordChange": {
        "properties": {
          "current_password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Current Password"
          },
          "new_password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 10,
            "title": "New Password",
            "description": "至少 10 位，不能与当前口令相同"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "current_password",
          "new_password"
        ],
        "title": "PasswordChange"
      },
      "Preferences": {
        "properties": {
          "theme_mode": {
            "type": "string",
            "enum": [
              "light",
              "dark",
              "system"
            ],
            "title": "Theme Mode"
          },
          "theme_preset": {
            "type": "string",
            "enum": [
              "aurora-blue",
              "arc-purple",
              "quantum-cyan",
              "pulse-green",
              "corona-gold",
              "molten-orange",
              "rose-wave",
              "glacier-blue",
              "nebula-purple",
              "deep-space-gray",
              "neon-cyan",
              "custom"
            ],
            "title": "Theme Preset"
          },
          "custom_primary": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^#[0-9A-Fa-f]{6}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Custom Primary",
            "description": "theme_preset 为 custom 时的主色"
          },
          "table_columns": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "type": "object",
            "title": "Table Columns",
            "description": "按页面保存的列顺序与显隐，键为页面 id"
          },
          "revision": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Revision",
            "description": "客户端最后读到的版本号，用于并发冲突检测"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "theme_mode",
          "theme_preset",
          "revision"
        ],
        "title": "Preferences",
        "description": "账号级偏好；revision 为服务端当前版本号。"
      },
...
      "Principal": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Roles"
          },
          "permissions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Permissions"
          },
          "data_scope": {
            "$ref": "#/components/schemas/DataScope"
          },
          "must_change_password": {
            "type": "boolean",
            "title": "Must Change Password"
          },
          "visible_metrics": {
            "items": {
              "type": "string",
              "enum": [
                "cost",
                "revenue",
                "impressions",
                "clicks",
                "launches",
                "callbacks",
                "conversions",
                "roi",
                "ctr",
                "cpm",
                "cpc",
                "click_arpu",
                "launch_rate",
                "cvr",
                "conversion_cost",
                "callback_cost",
                "launch_callback_rate",
                "revenue_gap"
              ]
            },
            "type": "array",
            "title": "Visible Metrics",
            "description": "当前账号可查看的指标；前端据此裁剪列选择，接口仍独立校验"
          },
          "navigation": {
            "items": {
              "$ref": "#/components/schemas/NavGroup"
            },
            "type": "array",
            "title": "Navigation",
            "description": "按权限裁剪后的导航；仅用于展示，接口仍独立鉴权"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "user_id",
          "username",
          "display_name",
          "tenant_id",
          "roles",
          "permissions",
          "data_scope",
          "must_change_password",
          "visible_metrics",
          "navigation"
        ],
        "title": "Principal"
      },
...
      "RefreshRequest": {
        "properties": {
          "refresh_token": {
            "type": "string",
            "minLength": 1,
            "title": "Refresh Token"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "refresh_token"
        ],
        "title": "RefreshRequest"
      },
...
      "TokenPair": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token"
          },
          "token_type": {
            "type": "string",
            "const": "bearer",
            "title": "Token Type",
            "default": "bearer"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In",
            "description": "访问令牌有效秒数"
          },
          "must_change_password": {
            "type": "boolean",
            "title": "Must Change Password"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "access_token",
          "refresh_token",
          "expires_in",
          "must_change_password"
        ],
        "title": "TokenPair"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
  "servers": [
    {
      "url": "http://127.0.0.1:8080",
      "description": "本机 Java 业务服务",
      "x-environment": "development"
    },
    {
      "url": "http://127.0.0.1:18081",
      "description": "本机 Java 隔离验收示例，按需启动，非常驻入口",
      "x-environment": "testing"
    },
    {
      "url": "https://iaa-ops.invalid",
      "description": "公开演示规划占位，尚未部署；发布后替换为真实地址",
      "x-environment": "production"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
}